Fix: use courses&students Services in cart

+++ e-learning.API/Controllers/CartsController.cs
@@ -6,19 +6,27 @@ using Microsoft.AspNetCore.Mvc;
 public class CartController : ControllerBase
 {
     private readonly ICartService _cartService;
+    private readonly ICourseServices _courseServices;
+    private readonly IStudentServices _studentServices;
     private readonly IAuthenticationServices _authenticationServices;
 
 
-    public CartController(IAuthenticationServices authenticationServices, ICartService cartService)
+
+
+    public CartController(ICartService cartService, ICourseServices courseServices, IStudentServices studentServices, IAuthenticationServices authenticationServices)
     {
         _cartService = cartService;
+        _courseServices = courseServices;
+        _studentServices = studentServices;
         _authenticationServices = authenticationServices;
-
     }
 
     [HttpGet("{studentId}")]
     public async Task<IActionResult> GetCart(int studentId)
     {
+        var getStudent = await _studentServices.GetStudentAsync(studentId);
+        if (getStudent == null)
+            return NotFound("Student is not found !");
         var cart = await _cartService.GetCartAsync(studentId);
         return Ok(cart);
     }
@@ -26,6 +34,12 @@ public class CartController : ControllerBase
     [HttpPost("{studentId}/add")]
     public async Task<IActionResult> AddToCart([FromHeader] string token, int studentId, int courseId)
     {
+        var getStudent = await _studentServices.GetStudentAsync(studentId);
+        if (getStudent == null)
+            return NotFound("Student is not found !");
+        var course = await _courseServices.GetCourseByIdAsync(courseId);
+        if (course == null)
+            return NotFound($"No course with this ID:{courseId} !");
 
         var accessToken = await _authenticationServices.ValidateToken(token);
         switch (accessToken)
@@ -45,7 +59,30 @@ public class CartController : ControllerBase
     [HttpDelete("{studentId}/remove/{courseId}")]
     public async Task<IActionResult> RemoveFromCart(int studentId, int courseId)
     {
+        var getStudent = await _studentServices.GetStudentAsync(studentId);
+        if (getStudent == null)
+            return NotFound("Student is not found !");
+        var course = await _courseServices.GetCourseByIdAsync(courseId);
+        if (course == null)
+            return NotFound($"No course with this ID:{courseId} !");
         await _cartService.RemoveFromCartAsync(studentId, courseId);
         return Ok("Delete is successfully !+");
     }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 }
+
